+++ src/project1/src/project1/Ticket.java
@@ -20,7 +20,7 @@ public class Ticket implements Matchable<String>, Serializable {
     private Calendar dateOfPlay;
     private int ticketType;
     protected Play ticketPlay;
-    private int ticketPrice;
+    private long ticketPrice;
             
     public Ticket (int type, String ID, String serial, List<Credit> credit, Calendar calandar, Play aPlay) {
         this.customerID = ID;
@@ -51,7 +51,7 @@ public class Ticket implements Matchable<String>, Serializable {
      * 
      * @return price of this ticket
      */
-    public int getPrice () {
+    public long getPrice () {
         return ticketPrice;
     }
     
@@ -77,7 +77,7 @@ public class Ticket implements Matchable<String>, Serializable {
      * @param aPlay
      * @return the ticket price depending on ticket type
      */
-    private  int setTicketPrice (int type, Play aPlay) {
+    private  long setTicketPrice (int type, Play aPlay) {
             switch (type) {
                 case Theater.REGULARTICKET:
                     return  RegularTicket.ticketPrice(aPlay);
